Update run.sh
Modified run.sh to have a proper first line which configures the terminal (or whatever that thing does!)
Also removed the Python plotting script; the user can run this separately, it is a failure point. 
I have also removed the compilation step, so the user can do it separately. Hopefully this makes it more robust to failure.